New Version 1.7.0 with webinterface
