+++ rpi1wire/_pv_1_6_0/plugin.yaml
@@ -14,7 +14,7 @@ plugin:
     documentation: https://github.com/smarthomeNG/plugins/rpi1wire/Readme.md        # url of documentation (wiki) page
     support: https://knx-user-forum.de/forum/supportforen/smarthome-py/1451916-support-thread-für-rpi1wire-plugin
 
-    version: 1.7.0                  # Plugin version
+    version: 1.6.0                  # Plugin version
     sh_minversion: 1.5              # minimum shNG version to use this plugin
 #    sh_maxversion:                 # maximum shNG version to use this plugin (leave empty if latest)
     multi_instance: False           # plugin supports multi instance
@@ -46,26 +46,11 @@ item_attributes:
             en: 'Sensor address in xx-yyyyyyyyyyyy format (Please note: If the sensor address is numerical, it has to be enclosed in quotes) or an alias'
             fr: 'Adresse du capteur au format xx-yyyyyyyyyyyy'
 
-    # Alternativ rpi1wire:
-    rpi1wire_name:
-    rpi1wire_id:
-        type: str
-        description:
-            de: 'Alternativ zu Sensoradresse kann auch ein Name in der Form rpi-temp1 vergeben werden, wobei die Zahl am Ende die laufende Nummer der gefundenen Sensoren bestimmt.'
-            en: 'As an alternative to the sensor address, a name in the form rpi-temp1 can also be assigned, with the number at the end determining the serial number of the sensors found.'
-         
     rpi1wire_update:
         type: bool
         description:
             de: 'Started bei Triggerung ein Update der Sensorliste'
             en: 'Rereads the sensor list after being triggered'
-
-    rpi1wire_sys:
-        type: str
-        description:
-            de: 'Mögliche Werte sind count oder list. Bei count wird die Anzahl der gefundenen Sensoren dem item.value als Integer übergeben, bei list wird eine Liste der gefundenen Sensoren als String übergeben'
-            en: 'Possible values ​​are count or list. With count the number of sensors found is transferred to item.value as an integer, with list a list of the sensors found is transferred as a string'
-
 
 item_structs: NONE
     # Definition of item-structure templates for this plugin (enter 'item_structs: NONE', if section should be empty)
