+++ rpi1wire/_pv_1_6_0/webif/static/img/readme.txt
@@ -1,0 +1,6 @@
+This directory is for storing images that are used by the web interface.
+
+If you want to have your own logo on the top of the web interface, store it here and name it plugin_logo.<extension>.
+
+Extension can be png, svg or jpg
+
