+++ rpi1wire/webif/templates/index.html
@@ -30,26 +30,26 @@
 <table class="table table-striped table-hover">
 	<tbody>
 		<tr>
-			<td class="py-1"><strong>Prompt 1</strong></td>
-			<td class="py-1">{% if 1 == 2 %}{{ _('Ja') }}{% else %}{{ _('Nein') }}{% endif %}</td>
+			<td class="py-1"><strong>Class</strong></td>
+			<td class="py-1">{{ classname }}</td>
 			<td class="py-1" width="50px"></td>
-			<td class="py-1"><strong>Prompt 4</strong></td>
-			<td class="py-1">{{ _('Wert 4') }}</td>
+			<td class="py-1"><strong>-</strong></td>
+			<td class="py-1">-</td>
 			<td class="py-1" width="50px"></td>
 		</tr>
 		<tr>
-			<td class="py-1"><strong>Prompt 2</strong></td>
-			<td class="py-1">{{ _('Wert 2') }}</td>
+			<td class="py-1"><strong>Cycle</strong></td>
+			<td class="py-1">{{ cycle }}</td>
 			<td></td>
-			<td class="py-1"><strong>Prompt 5</strong></td>
+			<td class="py-1"><strong>-</strong></td>
 			<td class="py-1">-</td>
 			<td></td>
 		</tr>
 		<tr>
-			<td class="py-1"><strong>Prompt 3</strong></td>
-			<td class="py-1">-</td>
+			<td class="py-1"><strong>Pfad</strong></td>
+			<td class="py-1">{{ dirname }}</td>
 			<td></td>
-			<td class="py-1"><strong>Prompt 6</strong></td>
+			<td class="py-1"><strong>-</strong></td>
 			<td class="py-1">-</td>
 			<td></td>
 		</tr>
@@ -72,7 +72,7 @@
 <!--
 	Define the number of tabs for the body of the web interface (1 - 3)
 -->
-{% set tabcount = 4 %}
+{% set tabcount = 2 %}
 
 
 <!--
@@ -88,8 +88,29 @@
 -->
 {% set tab1title = "<strong>" ~ p.get_shortname() ~ " Items</strong> (" ~ item_count ~ ")" %}
 {% block bodytab1 %}
-<div class="container-fluid m-2">
-	{{ _('Hier kommt der Inhalt des Webinterfaces hin.') }}
+<div class="table-responsive" style="margin-left: 3px; margin-right: 3px;" class="row">
+    <div class="col-sm-12">
+        <table class="table table-striped table-hover pluginList">
+            <thead>
+                <tr>
+                    <th>{{ _('Item') }}</th>
+                    <th>{{ _('Typ') }}</th>
+                    <th>{{ _('Wert') }}</th>
+                    <th>{{ _('Aktualisiert') }}</th>
+                </tr>
+            </thead>
+            <tbody>
+                {% for item in items %}
+                    <tr>
+                        <td class="py-1">{{ item._path }}</td>
+                        <td class="py-1">{{ item._type }}</td>
+                        <td class="py-1">{{ item._value }}</td>
+                        <td class="py-1">{{ item.property.last_update }}</td>
+                    </tr>
+                {% endfor %}
+            </tbody>
+        </table>
+    </div>
 </div>
 {% endblock bodytab1 %}
 
@@ -97,8 +118,29 @@
 <!--
 	Content block for the second tab of the Webinterface
 -->
-{% set tab2title = "<strong>" ~ p.get_shortname() ~ " Geräte</strong> (" ~ device_count ~ ")" %}
+{% set tab2title = "<strong>" ~ p.get_shortname() ~ " Sensoren</strong> (" ~ device_count ~ ")" %}
 {% block bodytab2 %}
+<div class="table-responsive" style="margin-left: 3px; margin-right: 3px;" class="row">
+    <div class="col-sm-12">
+        <table class="table table-striped table-hover pluginList">
+            <thead>
+                <tr>
+                    <th>{{ _('Name') }}</th>
+                    <th>{{ _('id') }}</th>
+                </tr>
+            </thead>
+            <tbody>
+                {% for name in sensors %}
+                    <tr>
+                        <td class="py-1">{{ name }}</td>
+                        <td class="py-1">{{ sensors[name] }}</td>
+                    </tr>
+                {% endfor %}
+            </tbody>
+        </table>
+    </div>
+</div>
+
 {% endblock bodytab2 %}
 
 
@@ -108,8 +150,30 @@
 		{% set tab3title = "<strong>" ~ p.get_shortname() ~ " Geräte</strong>" %}
 
 	It has to be defined before (and outside) the block bodytab3
--->
+
 {% block bodytab3 %}
+<!--<div class="table-responsive" style="margin-left: 3px; margin-right: 3px;" class="row">
+    <div class="col-sm-12">
+        <table class="table table-striped table-hover pluginList">
+            <thead>
+                <tr>
+                    <th>{{ _('Item') }}</th>
+                    <th>{{ _('Typ') }}</th>
+                    <th>{{ _('Value') }}</th>
+                </tr>
+            </thead>
+            <tbody>
+                {% for value in values %}
+                    <tr>
+                        <td class="py-1">{{ value }}</td>
+                        <td class="py-1">{{ value }}</td>
+                        <td class="py-1">{{ value }}</td>
+                    </tr>
+                {% endfor %}
+            </tbody>
+        </table>
+    </div>
+</div>
 {% endblock bodytab3 %}
 
 
@@ -119,6 +183,7 @@
 		{% set tab4title = "<strong>" ~ p.get_shortname() ~ " Geräte</strong>" %}
 
 	It has to be defined before (and outside) the block bodytab4
--->
+
 {% block bodytab4 %}
 {% endblock bodytab4 %}
+-->
